fix: json schema validation and error modal
