+++ package.json
@@ -30,6 +30,7 @@
     "electron-is-dev": "^1.2.0",
     "electron-reload": "^1.5.0",
     "eslint": "^7.28.0",
+    "jsonschema": "^1.4.0",
     "lodash": "^4.17.21",
     "mapbox-gl": "^2.3.1",
     "react": "^17.0.1",
