+++ src/components/Modal/FileErrorModal/index.ts
@@ -1,0 +1,3 @@
+import FileErrorModal from './FileErrorModal';
+
+export default FileErrorModal;
