+++ src/components/Modal/index.ts
@@ -1,0 +1,3 @@
+import Modal from './Modal';
+
+export default Modal;
